requirements.txt
boto3>=1.34.0
botocore>=1.34.0
datadog-api-client>=2.0.0
numpy>=1.26.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
requests>=2.28.0
streamlit>=1.28.0
